+++ src/snowflake/snowpark/dataframe.py
@@ -710,7 +710,7 @@ class DataFrame:
         )
 
     def __copy__(self) -> "DataFrame":
-        return DataFrame(self._session, copy.copy(self._plan))
+        return DataFrame(self._session, copy.copy(self._select_statement or self._plan))
 
     if installed_pandas:
         import pandas  # pragma: no cover
@@ -3824,6 +3824,15 @@ Query List:
         exprs = [convert(col) for col in parse_positional_args_to_list(*cols)]
         return exprs
 
+    def print_schema(self) -> None:
+        schema_tmp_str = "\n".join(
+            [
+                f" |-- {attr.name}: {attr.datatype} (nullable = {str(attr.nullable)})"
+                for attr in self._plan.attributes
+            ]
+        )
+        print(f"root\n{schema_tmp_str}")
+
     where = filter
 
     # Add the following lines so API docs have them
@@ -3855,6 +3864,7 @@ Query List:
     randomSplit = random_split
     order_by = sort
     orderBy = order_by
+    printSchema = print_schema
 
     # These methods are not needed for code migration. So no aliases for them.
     # groupByGrouping_sets = group_by_grouping_sets