+++ CHANGELOG.md
@@ -4,11 +4,26 @@
 
 ### New Features
 
+- Added support for VOLATILE/IMMUTABLE keyword when registering UDFs.
+- Added support for specifying clustering keys when saving dataframes using `DataFrame.save_as_table`.
 - Accept `Iterable` objects input for `schema` when creating dataframes using `Session.create_dataframe`.
 - Added the property `DataFrame.session` to return a session object.
 - Added the property `Session.session_id` to return an integer that represents session id.
 - Added the property `Session.connection` to return a ServerConnection object .
 
+- Added support for creating a Snowpark session from a configuration file or environment variables.
+
+### Dependency updates
+
+- Updated ``snowflake-connector-python`` to 3.2.0.
+
+### Bug Fixes
+
+- Fixed a bug where automatic package upload would raise `ValueError` even when compatible package version were added in `session.add_packages`.
+- Fixed a bug where table stored procedures were not registered correctly when using `register_from_file`.
+- Fixed a bug where dataframe joins failed with `invalid_identifier` error.
+- Fixed a bug where `DataFrame.copy` disables SQL simplfier for the returned copy.
+- Fixed a bug where `session.sql().select()` would fail if any parameters are specified to `session.sql()`
 
 ## 1.7.0 (2023-08-28)